requirements.txt
torch
torchvision
torchaudio
openai-whisper
tqdm
fastapi
uvicorn
python-multipart
